+++ components/solid/panda.config.ts
@@ -7,4 +7,9 @@ export default defineConfig({
   include: ['./src/**/*.{js,jsx,ts,tsx}'],
   jsxFramework: 'solid',
   outdir: 'styled-system',
+  staticCss: {
+    recipes: {
+      input: ['*'],
+    },
+  },
 })